+++ test/validate.js
@@ -1,0 +1,7 @@
+import isEmpty from 'lodash/isEmpty'
+
+export function isRequired(value) {
+  if (isEmpty(value)) {
+    return 'Required'
+  }
+}